refactor：提取关注相关命令     - 将所有与“关注问题”相关的命令（如添加到关注、移除关注、置顶）从 `extension.ts` 提取到一个新的独立文件 `src/commands/focusCommands.ts` 中。     - 此举旨在降低 `extension.ts` 的复杂性，提高代码的可维护性和模块化程度。
